+++ src/components/status/status.js
@@ -1,4 +1,5 @@
 import ReplyButton from '../reply_button/reply_button.vue'
+import BookmarkButton from '../bookmark_button/bookmark_button.vue'
 import FavoriteButton from '../favorite_button/favorite_button.vue'
 import ReactButton from '../react_button/react_button.vue'
 import RetweetButton from '../retweet_button/retweet_button.vue'
@@ -55,6 +56,7 @@ library.add(
 const Status = {
   name: 'Status',
   components: {
+    BookmarkButton,
     ReplyButton,
     FavoriteButton,
     ReactButton,